pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bmad-orchestrator"
version = "0.1.0"
description = "BMAD Autonomous Engineering Orchestrator — LangGraph-powered end-to-end BMAD workflow automation"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langgraph-checkpoint-sqlite>=1.0.0",
    "anthropic>=0.40.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "jira>=3.8.0",
    "typer>=0.13.0",
    "rich>=13.9.0",
    "structlog>=24.4.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "questionary>=2.0.0",
    "claude-agent-sdk>=0.1.40",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
bmad-orchestrator = "bmad_orchestrator.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/bmad_orchestrator"]

[tool.hatch.build.targets.wheel.force-include]
".claude/commands" = "bmad_orchestrator/personas/data"

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-pyyaml>=6.0.12",
    "types-requests>=2.32.0",
    "debugpy>=1.8.20",
    "pytest-asyncio>=0.24.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=bmad_orchestrator --cov-report=term-missing --cov-fail-under=90"

[tool.coverage.run]
source = ["src/bmad_orchestrator"]
omit = ["*/tests/*", "*/cli.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "C4", "PIE"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.11"
strict = true
explicit_package_bases = true
mypy_path = "src"
